+++ NEWS.md
@@ -1,6 +1,6 @@
 # sgplot 0.2.0
 
-* Add [Scottish Government Design System colour palettes](https://designsystem.gov.scot/guidance/charts)
+* Add [Scottish Government Design System colour palettes](https://designsystem.gov.scot/guidance/charts/data-visualisation-colour-palettes)
 * Reduce `base_line_size` in `theme_sg()`
 * Remove default dark blue outline from `geom_col` and `geom_bar` when using `use_sgplot()`.